Refactor master stress bench
